routes update

+++ server.js
@@ -33,42 +33,45 @@ app.get('/login', function(req, res) {
 	res.render('login') ;
 }) ;
 
-app.post('/api/add', function(req, res){
-	if (req.body.message.length > 0) {
+app.post('/api/user/:userid/message/:action/:id?', function(req, res){
+	console.log("%s message(s)", req.params.action) ;
 
-		var mongoUrl = dbHost + req.body.userid;  
+	var mongoUrl = dbHost + req.params.userid;  
+
+	if (req.params.action == 'add')	{
+		if (req.body.message.length > 0) {
+
+			MongoClient.connect(mongoUrl, function(err, db) {
+				var collection = db.collection('messages') ;
+				collection.save({text: req.body.message}) ;
+				db.close() ;
+
+				res.redirect('/user/' + req.params.userid);
+			}) ;
+		}
+		else {
+				res.redirect('/user/' + req.params.userid);
+		}
+	} 
+  else if (req.params.action == "delete") {
+  	console.log('message id: %s', req.params.id) ;
+
+		MongoClient.connect(mongoUrl, function(err, db) {
+			db.collection('messages').remove({_id: mongodb.ObjectID(req.params.id)}) ;
+			db.close() ;
+			res.redirect('/user/' + req.params.userid);
+		}) ;
+
+	}
+	else if (req.params.action == "removeall") {
+		console.log("removeall") ;
 
 		MongoClient.connect(mongoUrl, function(err, db){
-			var collection = db.collection('messages') ;
-			collection.save({text: req.body.message}) ;
+			db.collection('messages').drop() ;
 			db.close() ;
 			res.redirect('/user/' + req.body.userid);
 		}) ;
 	}
-}) ;
-
-app.post('/api/clear', function(req, res){
-	console.log("clear") ;
-
-	var mongoUrl = dbHost + req.body.userid;  
-
-	MongoClient.connect(mongoUrl, function(err, db){
-		db.collection('messages').drop() ;
-		db.close() ;
-		res.redirect('/user/' + req.body.userid);
-	}) ;
-}) ;
-
-app.post('/api/user/:userid/message/:action/:id', function(req, res){
-	console.log("%s message: %s", req.params.action, req.params.id) ;
-
-	var mongoUrl = dbHost + req.params.userid;  
-
-	MongoClient.connect(mongoUrl, function(err, db){
-		db.collection('messages').remove({_id: mongodb.ObjectID(req.params.id)}) ;
-		db.close() ;
-		res.redirect('/user/' + req.params.userid);
-	}) ;
 }) ;
 
 app.listen(port, function(){
